fix : code-rabbit review 반영

+++ src/main/java/com/back/global/security/CustomAuthenticationFilter.java
@@ -121,7 +121,7 @@ public class CustomAuthenticationFilter extends OncePerRequestFilter {
         }
 
         UserDetails user = new SecurityUser(
-                member.getId().intValue(),
+                member.getId(),
                 member.getEmail(),
                 member.getName(),
                 member.getPassword(),
