+++ src/main/java/com/back/global/security/CustomUserDetailsService.java
@@ -19,7 +19,7 @@ public class CustomUserDetailsService implements UserDetailsService {
                 .orElseThrow(() -> new UsernameNotFoundException("사용자를 찾을 수 없습니다."));
 
         return new SecurityUser(
-                member.getId().intValue(),
+                member.getId(),
                 member.getEmail(),
                 "",
                 member.getName(),
